wip feature/composition Move files to typesetting folder
- Move files
- Rename namespace
